major refactoring python files and templates
