+++ src/main/python/starlake-airflow/ai/starlake/airflow/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ['airflow_starlake_job', 'airflow_starlake_options']
 
-from .airflow_starlake_job import AirflowStarlakeJob, DEFAULT_DAG_ARGS, DEFAULT_POOL
-from .airflow_starlake_options import AirflowStarlakeOptions
+from .starlake_airflow_job import StarlakeAirflowJob, DEFAULT_DAG_ARGS, DEFAULT_POOL
+from .starlake_airflow_options import StarlakeAirflowOptions
