+++ src/main/python/starlake-airflow/ai/starlake/airflow/bash/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ['airflow_starlake_bash_job']
 
-from .airflow_starlake_bash_job import AirflowStarlakeBashJob
+from .starlake_airflow_bash_job import StarlakeAirflowBashJob
