+++ src/main/python/starlake-airflow/ai/starlake/airflow/gcp/__init__.py
@@ -1,4 +1,4 @@
 __all__ = ['airflow_starlake_cloud_run_job', 'airflow_starlake_dataproc_job']
 
-from .airflow_starlake_cloud_run_job import AirflowStarlakeCloudRunJob
-from .airflow_starlake_dataproc_job import AirflowStarlakeDataprocJob, StarlakeDataprocCluster, StarlakeDataprocClusterConfig, StarlakeDataprocMasterConfig, StarlakeDataprocWorkerConfig
+from .starlake_airflow_cloud_run_job import StarlakeAirflowCloudRunJob
+from .starlake_airflow_dataproc_job import StarlakeAirflowDataprocJob, StarlakeDataprocCluster, StarlakeDataprocClusterConfig, StarlakeDataprocMasterConfig, StarlakeDataprocWorkerConfig
