+++ src/main/resources/templates/dags/__starlake_airflow_bash_job.py.j2
@@ -1,8 +1,8 @@
-{% include 'templates/dags/__common.py.j2' %}
-from ai.starlake.airflow.gcp import AirflowStarlakeCloudRunJob
+{% include 'templates/dags/__common_airflow.py.j2' %}
+from ai.starlake.airflow.bash import StarlakeAirflowBashJob
 
 import sys
 
 #optional variable jobs as a dict of all options to apply by job
 #eg jobs = {"task1 domain.task1 name": {"options": "task1 transform options"}, "task2 domain.task2 name": {"options": "task2 transform options"}}
-sl_job = AirflowStarlakeCloudRunJob(options=dict(options, **sys.modules[__name__].__dict__.get('jobs', {})))
+sl_job = StarlakeAirflowBashJob(options=dict(options, **sys.modules[__name__].__dict__.get('jobs', {})))
