+++ src/main/resources/templates/dags/__starlake_dagster_shell_job.py.j2
@@ -1,8 +1,8 @@
 {% include 'templates/dags/__common_dagster.py.j2' %}
-from ai.starlake.dagster.shell import DagsterStarlakeShellJob
+from ai.starlake.dagster.shell import StarlakeDagsterShellJob
 
 import sys
 
 #optional variable jobs as a dict of all options to apply by job
 #eg jobs = {"task1 domain.task1 name": {"options": "task1 transform options"}, "task2 domain.task2 name": {"options": "task2 transform options"}}
-sl_job = DagsterStarlakeShellJob(options=dict(options, **sys.modules[__name__].__dict__.get('jobs', {})))
+sl_job = StarlakeDagsterShellJob(options=dict(options, **sys.modules[__name__].__dict__.get('jobs', {})))
