+++ src/main/resources/templates/dags/load/airflow_scheduled_table_dataproc.py.j2
@@ -22,5 +22,5 @@
 # - incoming_path: the path to the incoming landing zone (if not specified, the default SL_ROOT/incoming path will be used) [OPTIONAL]
 # - tags: a list of tags to be applied to the dag [OPTIONAL]
 # Naming rule: scheduled or sensor, global or domain or table, cloudrun or bash or dataproc or serverless with free-text
-{% include 'templates/dags/__starlake_dataproc_job.py.j2' %}
-{% include 'templates/dags/load/__scheduled_table_tpl.py.j2' %}
+{% include 'templates/dags/__starlake_airflow_dataproc_job.py.j2' %}
+{% include 'templates/dags/load/__airflow_scheduled_table_tpl.py.j2' %}
